Moving the ctx cleanup action to request chain
This makes the cleanup be triggered before request finalizer handler.

When client term the connection, nginx receives a writing error.
Then the request began to finish. But in this case, nginx setup a
ngx_http_request_finalizer to the write_event_handler, that's surely
an issue of nginx. It causes the request cannot be clean because a
broken connection cannot trigger write availd event anymore.

I move the cleanup action a bit earily to mitigate the issue. It
should be fixed in nginx.